Save,Update,deleteControllers

+++ Shopy/src/main/java/Ecommerce/example/Shopy/Service/CategoryService.java
@@ -1,39 +1,39 @@
 package Ecommerce.example.Shopy.Service;
 
+import Ecommerce.example.Shopy.Config.GlobalRESTAPIHandler;
 import Ecommerce.example.Shopy.Config.Response;
 import Ecommerce.example.Shopy.Entity.Category;
-import Ecommerce.example.Shopy.Entity.User;
 import Ecommerce.example.Shopy.Repository.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 public class CategoryService {
     @Autowired
     CategoryRepository categoryRepository;
 
-    public ResponseEntity<Response<Category>> createCategory(Category category) {
-        categoryRepository.save(category);
-        // Create a custom response
-        Response<Category> response = new Response<>(
-                "Created Successfully!",
-                HttpStatus.CREATED.value(),
-                category
-        );
-        // Return the response wrapped in ResponseEntity
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    @Autowired
+    GlobalRESTAPIHandler globalRESTAPIHandler;
+
+
+    public ResponseEntity<Response<Category>> saveOrUpdateCategory(Category category, String successMessage) {
+        Category savedCategory = categoryRepository.save(category);
+        return globalRESTAPIHandler.createResponse(successMessage, HttpStatus.OK, savedCategory);
     }
 
-    public ResponseEntity<Response<Category>> deleteCategory(Category category) {
-        categoryRepository.deleteById(category.getId());
-        Response<Category> response = new Response<>(
-                "Deleted Successfully!",
-                HttpStatus.CREATED.value(),
-                category
-        );
-        // Return the response wrapped in ResponseEntity
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    public ResponseEntity<Response<String>> deleteCategory(Category category) {
+        Optional<Category> categoryFound = categoryRepository.findById(category.getId());
+
+        if (categoryFound.isPresent()) {
+            categoryRepository.deleteById(category.getId());
+            return globalRESTAPIHandler.createResponse("Deleted Successfully!", HttpStatus.OK, "Cart ID: "
+                    + category.getId());
+        } else {
+            return globalRESTAPIHandler.createResponse("Category not found!", HttpStatus.NOT_FOUND, null);
+        }
     }
 }
