+++ Shopy/src/main/java/Ecommerce/example/Shopy/Service/ProductService.java
@@ -1,39 +1,38 @@
 package Ecommerce.example.Shopy.Service;
 
+import Ecommerce.example.Shopy.Config.GlobalRESTAPIHandler;
 import Ecommerce.example.Shopy.Config.Response;
 import Ecommerce.example.Shopy.Entity.Product;
-import Ecommerce.example.Shopy.Entity.Product_Photo;
 import Ecommerce.example.Shopy.Repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 public class ProductService {
     @Autowired
     ProductRepository productRepository;
 
-    public ResponseEntity<Response<Product>> createProduct(Product product) {
-        productRepository.save(product);
-        // Create a custom response
-        Response<Product> response = new Response<>(
-                "Created Successfully!",
-                HttpStatus.CREATED.value(),
-                product
-        );
-        // Return the response wrapped in ResponseEntity
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    @Autowired
+    GlobalRESTAPIHandler globalRESTAPIHandler;
+
+    public ResponseEntity<Response<Product>> saveOrUpdateProduct(Product product, String successMessage) {
+        Product savedProduct = productRepository.save(product);
+        return globalRESTAPIHandler.createResponse(successMessage, HttpStatus.OK, savedProduct);
     }
 
-    public ResponseEntity<Response<Product>> deleteProduct(Product product) {
-        productRepository.deleteById(product.getId());
-        Response<Product> response = new Response<>(
-                "Deleted Successfully!",
-                HttpStatus.CREATED.value(),
-                product
-        );
-        // Return the response wrapped in ResponseEntity
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    public ResponseEntity<Response<String>> deleteProduct(Product product) {
+        Optional<Product> productFound = productRepository.findById(product.getId());
+
+        if (productFound.isPresent()) {
+            productRepository.deleteById(product.getId());
+            return globalRESTAPIHandler.createResponse("Deleted Successfully!", HttpStatus.OK, "Cart ID: "
+                    + product.getId());
+        } else {
+            return globalRESTAPIHandler.createResponse("Product not found!", HttpStatus.NOT_FOUND, null);
+        }
     }
 }
