+++ Shopy/src/main/java/Ecommerce/example/Shopy/Service/Product_PhotoService.java
@@ -1,5 +1,6 @@
 package Ecommerce.example.Shopy.Service;
 
+import Ecommerce.example.Shopy.Config.GlobalRESTAPIHandler;
 import Ecommerce.example.Shopy.Config.Response;
 import Ecommerce.example.Shopy.Entity.Product_Photo;
 import Ecommerce.example.Shopy.Repository.Product_PhotoRepository;
@@ -8,29 +9,32 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 public class Product_PhotoService {
     @Autowired
     Product_PhotoRepository productPhotoRepository;
-        public ResponseEntity<Response<Product_Photo>> createPhoto_product(Product_Photo productPhoto) {
-            productPhotoRepository.save(productPhoto);
-        // Create a custom response
-        Response<Product_Photo> response = new Response<>(
-                "Created Successfully!",
-                HttpStatus.CREATED.value(),
-                productPhoto
-        );
-        // Return the response wrapped in ResponseEntity
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    @Autowired
+    GlobalRESTAPIHandler globalRESTAPIHandler;
+
+
+    public ResponseEntity<Response<Product_Photo>> saveOrUpdateProductPhoto(Product_Photo productPhoto, String
+            successMessage) {
+        Product_Photo savedProductPhoto = productPhotoRepository.save(productPhoto);
+        return globalRESTAPIHandler.createResponse(successMessage, HttpStatus.OK, savedProductPhoto);
     }
-    public ResponseEntity<Response<Product_Photo>> deletePhoto_product(Product_Photo productPhoto) {
-        productPhotoRepository.deleteById(productPhoto.getId());
-        Response<Product_Photo> response = new Response<>(
-                "Deleted Successfully!",
-                HttpStatus.CREATED.value(),
-                productPhoto
-        );
-        // Return the response wrapped in ResponseEntity
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+
+    public ResponseEntity<Response<String>> deleteProductPhoto(Product_Photo productPhoto) {
+        Optional<Product_Photo> productPhotoFound = productPhotoRepository.findById(productPhoto.getId());
+
+        if (productPhotoFound.isPresent()) {
+            productPhotoRepository.deleteById(productPhoto.getId());
+            return globalRESTAPIHandler.createResponse("Deleted Successfully!", HttpStatus.OK, "Cart ID: "
+                    + productPhoto.getId());
+        } else {
+            return globalRESTAPIHandler.createResponse("Product's photo  not found!", HttpStatus.NOT_FOUND,
+                    null);
+        }
     }
 }
