+++ Shopy/src/main/java/Ecommerce/example/Shopy/Service/ReviewService.java
@@ -1,5 +1,6 @@
 package Ecommerce.example.Shopy.Service;
 
+import Ecommerce.example.Shopy.Config.GlobalRESTAPIHandler;
 import Ecommerce.example.Shopy.Config.Response;
 import Ecommerce.example.Shopy.Entity.Review;
 import Ecommerce.example.Shopy.Repository.ReviewRepository;
@@ -8,34 +9,32 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 public class ReviewService {
 
     @Autowired
     private ReviewRepository reviewRepository;
 
-    public ResponseEntity<Response<Review>> createReview(Review review) {
-        // Save the review entity
-        reviewRepository.save(review);
-        // Create a custom response
-        Response<Review> response = new Response<>(
-                "Created Successfully!",
-                HttpStatus.CREATED.value(),
-                review
-        );
-        // Return the response wrapped in ResponseEntity
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    @Autowired
+    GlobalRESTAPIHandler globalRESTAPIHandler;
+
+    public ResponseEntity<Response<Review>> saveOrUpdateReview(Review review, String successMessage) {
+        Review savedReview = reviewRepository.save(review);
+        return globalRESTAPIHandler.createResponse(successMessage, HttpStatus.OK, savedReview);
     }
 
-    public ResponseEntity<Response<Review>> deleteReview(Review review) {
-        // delete the review entity
-        reviewRepository.deleteById(review.getId());
-        Response<Review> response = new Response<>(
-                "Deleted Successfully!",
-                HttpStatus.ACCEPTED.value(),
-                review
-        );
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    public ResponseEntity<Response<String>> deleteReview(Review review) {
+        Optional<Review> reviewFound = reviewRepository.findById(review.getId());
+
+        if (reviewFound.isPresent()) {
+            reviewRepository.deleteById(review.getId());
+            return globalRESTAPIHandler.createResponse("Deleted Successfully!", HttpStatus.OK, "Cart ID: "
+                    + review.getId());
+        } else {
+            return globalRESTAPIHandler.createResponse("Review not found!", HttpStatus.NOT_FOUND, null);
+        }
     }
