+++ Shopy/src/main/java/Ecommerce/example/Shopy/Service/SubCategoryService.java
@@ -1,7 +1,7 @@
 package Ecommerce.example.Shopy.Service;
 
+import Ecommerce.example.Shopy.Config.GlobalRESTAPIHandler;
 import Ecommerce.example.Shopy.Config.Response;
-import Ecommerce.example.Shopy.Entity.Review;
 import Ecommerce.example.Shopy.Entity.SubCategory;
 import Ecommerce.example.Shopy.Repository.SubCategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,33 +9,33 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 public class SubCategoryService {
 
     @Autowired
     SubCategoryRepository subCategoryRepository;
 
-    public ResponseEntity<Response<SubCategory>> createSubCategory(SubCategory subCategory) {
-        // Save the review entity
-        subCategoryRepository.save(subCategory);
-        // Create a custom response
-        Response<SubCategory> response = new Response<>(
-                "Created Successfully!",
-                HttpStatus.CREATED.value(),
-                subCategory
-        );
-        // Return the response wrapped in ResponseEntity
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    @Autowired
+    GlobalRESTAPIHandler globalRESTAPIHandler;
+
+    public ResponseEntity<Response<SubCategory>> saveOrUpdateSubCategory(SubCategory subCategory, String successMessage)
+    {
+        SubCategory savedSubCategory = subCategoryRepository.save(subCategory);
+        return globalRESTAPIHandler.createResponse(successMessage, HttpStatus.OK, subCategory);
     }
 
-    public ResponseEntity<Response<SubCategory>> deleteSubCategory(SubCategory subCategory) {
-        // delete the review entity
-        subCategoryRepository.deleteById(subCategory.getId());
-        Response<SubCategory> response = new Response<>(
-                "Deleted Successfully!",
-                HttpStatus.ACCEPTED.value(),
-                subCategory
-        );
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    public ResponseEntity<Response<String>> deleteSubCategory(SubCategory subCategory) {
+        Optional<SubCategory> subCategoryFound = subCategoryRepository.findById(subCategory.getId());
+
+        if (subCategoryFound.isPresent()) {
+            subCategoryRepository.deleteById(subCategory.getId());
+            return globalRESTAPIHandler.createResponse("Deleted Successfully!", HttpStatus.OK, "Cart ID: "
+                    + subCategory.getId());
+        } else {
+            return globalRESTAPIHandler.createResponse("Subcategory not found!", HttpStatus.NOT_FOUND,
+                    null);
+        }
     }
 }
