+++ Shopy/src/main/java/Ecommerce/example/Shopy/Service/UserService.java
@@ -1,5 +1,6 @@
 package Ecommerce.example.Shopy.Service;
 
+import Ecommerce.example.Shopy.Config.GlobalRESTAPIHandler;
 import Ecommerce.example.Shopy.Config.Response;
 import Ecommerce.example.Shopy.Entity.User;
 import Ecommerce.example.Shopy.Repository.UserRepository;
@@ -8,33 +9,33 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 public class UserService {
 
     @Autowired
     UserRepository userRepository;
 
-    public ResponseEntity<Response<User>> createUser(User user) {
-        // Save the user entity
-        userRepository.save(user);
-        // Create a custom response
-        Response<User> response = new Response<>(
-                "Created Successfully!",
-                HttpStatus.CREATED.value(),
-                user
-        );
-        // Return the response wrapped in ResponseEntity
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    @Autowired
+    GlobalRESTAPIHandler globalRESTAPIHandler;
+
+    public ResponseEntity<Response<User>> saveOrUpdateUser(User user, String successMessage)
+    {
+        User savedUser = userRepository.save(user);
+        return globalRESTAPIHandler.createResponse(successMessage, HttpStatus.OK, user);
     }
-    public ResponseEntity<Response<User>> deleteUser(User user) {
-        userRepository.deleteById(user.getId());
-        Response<User> response = new Response<>(
-                "Deleted Successfully!",
-                HttpStatus.CREATED.value(),
-                user
-        );
-        // Return the response wrapped in ResponseEntity
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+
+    public ResponseEntity<Response<String>> deleteUser(User user) {
+        Optional<User> userFound = userRepository.findById(user.getId());
+
+        if (userFound.isPresent()) {
+            userRepository.deleteById(user.getId());
+            return globalRESTAPIHandler.createResponse("Deleted Successfully!", HttpStatus.OK, "Cart ID: "
+                    + user.getId());
+        } else {
+            return globalRESTAPIHandler.createResponse("User not found!", HttpStatus.NOT_FOUND, null);
+        }
     }
 
 }
